changed structure a little
